remove trb-loader
