+++ gemfiles/rails_7.0.gemfile.lock
@@ -1,7 +1,7 @@
 PATH
   remote: ..
   specs:
-    trailblazer-rails (2.2.0)
+    trailblazer-rails (2.3.0)
       railties (>= 5.2.0)
       trailblazer (>= 2.1.0, < 2.2.0)
