+++ lib/trailblazer/rails/version.rb
@@ -1,5 +1,5 @@
 module Trailblazer
   module Rails
-    VERSION = "2.2.0".freeze
+    VERSION = "2.3.0".freeze
   end
 end
