+++ test/dummy/app/concepts/params/operation/with_args.rb
@@ -1,5 +1,7 @@
 module Params
+  module Operation
   class WithArgs < Trailblazer::Operation
     step ->(context, **) { context[:x] = %{#{context[:params]} #{context[:current_user]}} }
+  end
   end
 end
