+++ test/dummy/app/concepts/params/operation/with_variables.rb
@@ -1,5 +1,7 @@
 module Params
+  module Operation
   class WithVariables < Trailblazer::Operation
     step ->(ctx, controller_name:, **) { ctx[:model] = controller_name }
+  end
   end
 end
