+++ test/dummy/app/controllers/args_controller.rb
@@ -1,6 +1,6 @@
 class ArgsController < ApplicationController
   def with_args
-    run Params::WithArgs, params: {fake: "bla"}
+    run Params::Operation::WithArgs, params: {fake: "bla"}
   end
 
   private
