+++ test/dummy/app/controllers/songs_controller.rb
@@ -22,7 +22,7 @@ class SongsController < ApplicationController
   end
 
   def with_variables
-    run Params::WithVariables, controller_name: self.class.to_s
+    run Params::Operation::WithVariables, controller_name: self.class.to_s
 
     render html: %{<h1>#{@model.inspect}</h1>}.html_safe
   end
